+++ util/src/main/java/com/sabbreview/controller/UserController.java
@@ -17,6 +17,12 @@ import java.util.List;
 import javax.persistence.RollbackException;
 import javax.persistence.TypedQuery;
 
+/**
+ * Contains the high level code for operations on User JPA Objects.
+ * Authentication is enforced for some methods in this class.
+ * Call this class to do operations on Users.
+ * @see User
+ */
 public class UserController extends Controller {
   private static final String EMAIL_REGEX;
 
@@ -24,6 +30,12 @@ public class UserController extends Controller {
     EMAIL_REGEX = "^\\w+([-+.']\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*$";
   }
 
+  /**
+   * Stores a user object in the database.
+   * ALL USERS ARE CREATED AS ADMINS CURRENTLY.
+   * SEE TODO!!!
+   * @param user User to persist.
+   */
   public static TransactionState<User> registerUser(User user) {
     try {
       em.getTransaction().begin();
@@ -39,9 +51,9 @@ public class UserController extends Controller {
       rollback();
       e.printStackTrace();
       return new TransactionState<>(null, TransactionStatus.STATUS_ERROR,
-          (e.getCause().getMessage().contains("duplicate") ?
-              "There is already a user with that email address" :
-              null));
+              (e.getCause().getMessage().contains("duplicate") ?
+                      "There is already a user with that email address" :
+                      null));
     } catch (Exception e) {
       rollback();
       e.printStackTrace();
@@ -49,6 +61,11 @@ public class UserController extends Controller {
     }
   }
 
+  /**
+   * Fetches a user object.
+   * @param emailAddress Email address of user to fetch.
+   * @return The user object, as part of a transactionstate.
+   */
   public static TransactionState<User> getUser(String emailAddress) {
     User user = em.find(User.class, emailAddress);
     if (user == null) {
@@ -58,6 +75,11 @@ public class UserController extends Controller {
     }
   }
 
+  /**
+   * Generates a session token.
+   * @param uap
+   * @return A token, as part of a transactionstate.
+   */
   public static TransactionState<Token> generateSession(UserAuthenticationParameters uap) {
     String token;
     try {
@@ -96,16 +118,25 @@ public class UserController extends Controller {
 
   }
 
+  /**
+   * Deletes the user that calls this function.
+   * @param principle Principle of the calling user.
+   */
   public static TransactionState<User> deleteUser(String principle) {
     try {
       em.getTransaction().begin();
       User user = em.find(User.class, principle);
-      if(user == null) {
-        throw new ValidationException();
-      } else {
-        em.remove(user);
+      if(user.getEmailAddress().equals(principle)) {
+
+        em.getTransaction().begin();
+        if (user == null) {
+          throw new ValidationException();
+        } else {
+          em.remove(user);
+        }
+        em.getTransaction().commit();
+
       }
-      em.getTransaction().commit();
     } catch (RollbackException e) {
       rollback();
       return new TransactionState<>(null, TransactionStatus.STATUS_ERROR, e.getMessage());
@@ -141,7 +172,7 @@ public class UserController extends Controller {
   public static TransactionState<List<Application>> getApplicationsForUser(String principle) {
     try {
       TypedQuery<Application> applicationTypedQuery = em.createNamedQuery("get-all-applications-for-user", Application.class);
-      applicationTypedQuery.setParameter("principle", principle);
+      applicationTypedQuery.setParameter("id", principle);
       List<Application> applicationList = applicationTypedQuery.getResultList();
       return new TransactionState<>(applicationList, TransactionStatus.STATUS_OK);
     } catch (Exception e) {
@@ -165,6 +196,11 @@ public class UserController extends Controller {
   }
 
 
+  /**
+   * Fetches all the assignments for the calling user.
+   * @param principle Principle of the calling user.
+   * @return A list of Assignments.
+   */
   public static TransactionState<List<Assignment>> getAssignmentsForUser(String principle) {
     try {
       TypedQuery<Assignment> assignmentTypedQuery = em.createNamedQuery("get-all-assignments-for-user", Assignment.class);
